# web/openapi/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT or API key",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "OpenAI-compatible AI control-plane gateway. This document covers the full HTTP surface; request/response bodies are summarized — see the admin UI for live examples.",
    "title": "AI Proxy Gateway API",
    "version": "v0.80.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/admin/apps": {
      "get": {
        "operationId": "get__admin_apps",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List/create AI work apps",
        "tags": [
          "apps"
        ]
      },
      "post": {
        "operationId": "post__admin_apps",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List/create AI work apps",
        "tags": [
          "apps"
        ]
      }
    },
...
    "/admin/apps/{id}": {
      "delete": {
        "operationId": "delete__admin_apps_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get/update/delete a work app",
        "tags": [
          "apps"
        ]
      },
      "get": {
        "operationId": "get__admin_apps_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get/update/delete a work app",
        "tags": [
          "apps"
        ]
      },
      "patch": {
        "operationId": "patch__admin_apps_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get/update/delete a work app",
        "tags": [
          "apps"
        ]
      },
      "put": {
        "operationId": "put__admin_apps_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get/update/delete a work app",
        "tags": [
          "apps"
        ]
      }
    },
...
    "/v1/apps": {
      "get": {
        "operationId": "get__v1_apps",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List AI work apps visible to the caller",
        "tags": [
          "apps"
        ]
      }
    },
    "/v1/apps/{id}": {
      "get": {
        "operationId": "get__v1_apps_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get one AI work app the caller may see",
        "tags": [
          "apps"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "this gateway",
      "url": "/"
    }
  ],
...
}